Generalize state manager helper structs.
Moves and changes the helper structs for RedisStateManager so they can
be reused by other StateManagers such as MemoryStateManager.
